+++ src.com.cod.mvc/controller/Observer.java
@@ -1,4 +1,6 @@
-package cod.mvc;
+package controller;
+
+import model.Coche;
 
 public interface Observer {
     void update(Coche arg);
