+++ src.com.cod.mvc/view/View.java
@@ -1,4 +1,6 @@
-package cod.mvc;
+package view;
+
+import model.Model;
 
 public class View {
     /**
@@ -7,6 +9,6 @@ public class View {
      * @param velocidad variable que se quiere mostrar
      */
     public static void muestraVelocidad(String matricula,int velocidad){
-        System.out.println("La velocidad del coche con matricula "+matricula+", es "+Model.getVelocidad(matricula));
+        System.out.println("La velocidad del coche con matricula "+matricula+", es "+ Model.getVelocidad(matricula));
     }
 }
