--- src/cod/mvc/App.java
@@ -1,7 +1,0 @@
-package cod.mvc;
-
-public class App {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
